FileAPI/FileReader_InvalidStateError.html

<!DOCTYPE html>
<title>FileAPI Test: FileReader InvalidStateError Exception</title>
<link rel="author" title="Yinkan Li" href="mailto:liyinkan.biz@gmail.com">
<link rel="help" href="http://www.w3.org/TR/FileAPI/#MultipleReads">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<div id="log"></div>
<script>


test(function() {
	var blob_1 = new Blob(['TEST000000001'])
	var blob_2 = new Blob(['TEST000000002'])
	var reader = new FileReader();
	assert_true(reader !== undefined && reader !== null);
	reader.readAsText(blob_1)
	assert_equals(reader.readyState, FileReader.LOADING, "readyState Must be LOADING")
	assert_throws("InvalidStateError", function () { 
		reader.readAsText(blob_2)
	}, "InvalidStateError DOM exception.")
}, 'test FileReader InvalidStateError exception for readAsText');

test(function() {
	var blob_1 = new Blob(['TEST000000001'])
	var blob_2 = new Blob(['TEST000000002'])
	var reader = new FileReader();
	assert_true(reader !== undefined && reader !== null);
	reader.readAsDataURL(blob_1)
	assert_equals(reader.readyState, FileReader.LOADING, "readyState Must be LOADING")
	assert_throws("InvalidStateError", function () { 
		reader.readAsDataURL(blob_2)
	}, "InvalidStateError DOM exception.")
}, 'test FileReader InvalidStateError exception for readAsDataURL');

test(function() {
	var blob_1 = new Blob(['TEST000000001'])
	var blob_2 = new Blob(['TEST000000002'])
	var reader = new FileReader();
	assert_true(reader !== undefined && reader !== null);
	reader.readAsArrayBuffer(blob_1)
	assert_equals(reader.readyState, FileReader.LOADING, "readyState Must be LOADING")
	assert_throws("InvalidStateError", function () { 
		reader.readAsArrayBuffer(blob_2)
	}, "InvalidStateError DOM exception.")
}, 'test FileReader InvalidStateError exception for readAsArrayBuffer');

test(function() {
	var blob_1 = new Blob(['TEST000000001'])
	var blob_2 = new Blob(['TEST000000002'])
	var reader = new FileReader();
	assert_true(reader !== undefined && reader !== null);
	reader.onloadstart = function() {
		assert_throws("InvalidStateError", function () { 
			reader.readAsArrayBuffer(blob_2)		
		}, "InvalidStateError DOM exception.")
	}
	reader.readAsArrayBuffer(blob_1)
	assert_equals(reader.readyState, FileReader.LOADING, "readyState Must be LOADING")
}, 'test FileReader InvalidStateError exception in onloadstart event for readAsArrayBuffer');

test(function() {
	var blob_1 = new Blob(['TEST000000001'])
	var blob_2 = new Blob(['TEST000000002'])
	var reader = new FileReader();
	assert_true(reader !== undefined && reader !== null);
	reader.onloadend = function(){
		reader.readAsArrayBuffer(blob_2)
		assert_equals(reader.readyState, FileReader.LOADING, "readyState Must be LOADING")
	}
	reader.readAsArrayBuffer(blob_1)
	assert_equals(reader.readyState, FileReader.LOADING, "readyState Must be LOADING")
}, 'test FileReader no InvalidStateError exception in onloadstart event for readAsArrayBuffer');

</script>
